+++ src/api/controllers/alerts/alerts.spec.js
@@ -2,13 +2,13 @@ import chai from 'chai';
 import sinon from 'sinon';
 import supertest from 'supertest';
 import bodyParser from 'body-parser';
+import express from 'express';
+import mockery from 'mockery';
+import queryStringPaging from '../../middleware/query-string-paging-middleware';
 
 import { FakeAuthMiddleware } from '../../test-helper';
-
-import express from 'express';
-import mockery from "mockery";
-
 import { AlertsProvider } from '../../../data/mongodb/index';
+import * as consts from '../../../../consts';
 
 chai.should();
 const expect = chai.expect;
@@ -33,31 +33,43 @@ describe('Alerts API endpoints', () => {
     Alerts = require('./index').default;
     app = express();
     app.use(bodyParser.json());
+    app.use(queryStringPaging());
     request = supertest(app);
-    alertProvider = AlertsProvider({});
+
+    const db = {
+      collection: () => { },
+    };
+
+    alertProvider = AlertsProvider(db);
   });
 
   it('should get alerts for a passed gateway', (done) => {
     Alerts(app, [FakeAuthMiddleware(['samplegw', 'testgw'])()], { alertProvider });
-    const stub = sinon.stub(alertProvider, 'getAlerts')
-      .returns(Promise.resolve([{
-          gateway: 'samplegw',
-          date: new Date(),
-          deviceId: '00:11:22:33:44:55:66',
-          message: 'an alert',
-          read: false,
-          id: '58c7b3e46b835bf90cfdffeb',
-        }]),
+    const stub = sinon.stub(alertProvider, 'getPagedAlerts')
+      .returns(Promise.resolve({
+          list: [{
+            gateway: 'samplegw',
+            date: new Date(),
+            deviceId: '00:11:22:33:44:55:66',
+            message: 'an alert',
+            read: false,
+            id: '58c7b3e46b835bf90cfdffeb',
+          }],
+          hasNext: true,
+          lastId: '58c7b3e46b835bf90cfdffeb',
+        }),
       );
 
     request
-      .get('/api/alerts/?gw=samplegw')
+      .get('/api/alerts/?gw=samplegw&pageSize=1')
       .expect(200, (err, res) => {
         if (err) {
           done(err);
         } else {
-          stub.calledWith(['samplegw'])
-            .should.be.true;
+          stub.calledWith(sinon.match({ gateways: ['samplegw'] }), {
+            pageSize: 1,
+            lastObjectId: undefined
+          }).should.be.true;
           done();
         }
       });
@@ -65,25 +77,64 @@ describe('Alerts API endpoints', () => {
 
   it('should get alerts for some gateways', (done) => {
     Alerts(app, [FakeAuthMiddleware(['samplegw', 'testgw'])()], { alertProvider });
-    const stub = sinon.stub(alertProvider, 'getAlerts')
-      .returns(Promise.resolve([{
-          gateway: 'samplegw',
-          date: new Date(),
-          deviceId: '00:11:22:33:44:55:66',
-          message: 'an alert',
-          read: false,
-          id: '58c7b3e46b835bf90cfdffeb',
-        }]),
+    const stub = sinon.stub(alertProvider, 'getPagedAlerts')
+      .returns(Promise.resolve({
+          list: [{
+            gateway: 'samplegw',
+            date: new Date(),
+            deviceId: '00:11:22:33:44:55:66',
+            message: 'an alert',
+            read: false,
+            id: '58c7b3e46b835bf90cfdffeb',
+          }],
+          hasNext: true,
+          lastId: '58c7b3e46b835bf90cfdffeb',
+        }),
       );
+
 
     request
       .get('/api/alerts/?gw=samplegw&gw=testgw')
       .expect(200, (err, res) => {
         if (err) {
           done(err);
+          stub.restore();
         } else {
-          stub.calledWith(['samplegw', 'testgw'])
+          stub.calledWith(sinon.match({ gateways: ['samplegw', 'testgw'] }))
             .should.be.true;
+          done();
+          stub.restore();
+        }
+      });
+  });
+
+  it('should get alerts filtered for read and message fields', () => {
+    Alerts(app, [FakeAuthMiddleware(['samplegw', 'testgw'])()], { alertProvider });
+    const stub = sinon.stub(alertProvider, 'getPagedAlerts')
+      .returns(Promise.resolve({
+          list: [{
+            gateway: 'samplegw',
+            date: new Date(),
+            deviceId: '00:11:22:33:44:55:66',
+            message: 'an alert',
+            read: false,
+            id: '58c7b3e46b835bf90cfdffeb',
+          }],
+          hasNext: true,
+          lastId: '58c7b3e46b835bf90cfdffeb',
+        }),
+      );
+
+    request
+      .get('/api/alerts/?gw=samplegw&pageSize=1&text=testSearch&read=true')
+      .expect(200, (err, res) => {
+        if (err) {
+          done(err);
+        } else {
+          stub.calledWith(sinon.match({ text: 'testSearch', read: true }), {
+            pageSize: 1,
+            lastObjectId: undefined
+          }).should.be.true;
           done();
         }
       });
@@ -91,8 +142,8 @@ describe('Alerts API endpoints', () => {
 
   it('should responds 204 for not owned gateways', (done) => {
     Alerts(app, [FakeAuthMiddleware(['samplegw', 'testgw'])()], { alertProvider });
-    const stub = sinon.stub(alertProvider, 'getAlerts')
-      .returns(Promise.resolve([]),
+    const stub = sinon.stub(alertProvider, 'getPagedAlerts')
+      .returns(Promise.resolve({ list: [], hasNext: false, lastId: 0 }),
       );
 
     request
@@ -106,8 +157,38 @@ describe('Alerts API endpoints', () => {
       });
   });
 
+  it('should use default pageSize if query pageSize is more than max', () => {
+    Alerts(app, [FakeAuthMiddleware(['samplegw', 'testgw'])()], { alertProvider });
+    const stub = sinon.stub(alertProvider, 'getPagedAlerts')
+      .returns(Promise.resolve({
+          list: [{
+            gateway: 'samplegw',
+            date: new Date(),
+            deviceId: '00:11:22:33:44:55:66',
+            message: 'an alert',
+            read: false,
+            id: '58c7b3e46b835bf90cfdffeb',
+          }],
+          hasNext: true,
+          lastId: '58c7b3e46b835bf90cfdffeb',
+        }),
+      );
+
+
+    request
+      .get('/api/alerts/?gw=samplegw&gw=testgw&pageSize=500')
+      .expect(200, (err, res) => {
+        if (err) {
+          done(err);
+        } else {
+          stub.calledWith(['samplegw', 'testgw'], consts.PAGING_MAX_PAGE_SIZE, undefined).should.be.true;
+          done();
+        }
+      });
+  });
+
   it('should toggle read field', (done) => {
-    const stub = sinon.stub( alertProvider, 'getAlertById')
+    const stub = sinon.stub(alertProvider, 'getAlertById')
       .returns(Promise.resolve({
         gateway: 'samplegw',
         date: new Date(),
@@ -135,8 +216,8 @@ describe('Alerts API endpoints', () => {
       });
   });
 
-  it('should delete alert identifyed by id', (done) => {
-    const stubGet = sinon.stub( alertProvider, 'getAlertById')
+  it('should delete alert identified by id', (done) => {
+    const stubGet = sinon.stub(alertProvider, 'getAlertById')
       .returns(Promise.resolve({
         gateway: 'samplegw',
         date: new Date(),
@@ -163,7 +244,7 @@ describe('Alerts API endpoints', () => {
   });
 
   it('should return 403 trying to delete an alert of a wrong gateway', (done) => {
-    const stubGet = sinon.stub( alertProvider, 'getAlertById')
+    const stubGet = sinon.stub(alertProvider, 'getAlertById')
       .returns(Promise.resolve({
         gateway: 'test',
         date: new Date(),