+++ src/data/mongodb/events/eventsProvider.e2e.js
@@ -7,7 +7,7 @@ import EventsProvider from './eventsProvider';
 chai.should();
 const expect = chai.expect;
 
-describe('eventsProvider', () => {
+describe('eventProvider', () => {
   let subject;
   let db;
 
@@ -63,6 +63,58 @@ describe('eventsProvider', () => {
       })
       .catch(err => done(err));
 
+  });
+
+  it('should get last event by gateway, type and deviceId', (done) => {
+    subject = EventsProvider(db);
+
+    subject
+      .add({
+        GatewayId: 'VG59',
+        Type: 'TEST',
+        Payload: {
+          DeviceId: 'deviceid1',
+          Time: new Date(),
+          created: new Date(),
+          value: '5',
+        },
+      })
+      .then(() => {
+        subject
+          .add({
+            GatewayId: 'GTW',
+            Type: 'TEST1',
+            Payload: {
+              deviceId: 'deviceid',
+              Time: new Date(),
+              created: new Date(),
+            },
+          });
+      })
+      .then(() => {
+        subject
+          .add({
+            GatewayId: 'VG59',
+            Type: 'TEST1',
+            Payload: {
+              deviceId: 'deviceid1',
+              Time: new Date(),
+              created: new Date(),
+              value: '1',
+            },
+          });
+      })
+      .then(() => {
+        return subject.getLastEvent('VG59', ['TEST', 'TEST1'], 'deviceid1');
+      })
+      .then((res) => {
+        expect(res.GatewayId).to.equal('VG59');
+        expect(res.Type).to.equal('TEST1');
+        expect(res.Payload.deviceId).to.equal('deviceid1');
+        expect(res.Payload.value).to.equal('1');
+        done();
+      })
+      .catch(err => done(err));
   });
 
 });