+++ src/events/handlers/internal/api/powerStateHandler.js
@@ -1,4 +1,5 @@
 import logger from '../../../../common/logger';
+import PayloadFactory from '../../factory/payloadFactory';
 
 /**
  * Process the power switch request (typically coming from an API call).
@@ -8,6 +9,8 @@ import logger from '../../../../common/logger';
  * @param pnub
  * @constructor
  */
+
+const factory = new PayloadFactory();
 const PowerStateHandler = (deviceProvider, pnub) => ({
   process: ({ deviceId, gateway, param }) => {
     logger.log('info', `power action processor ${JSON.stringify({ deviceId, gateway, param })}`);
@@ -19,10 +22,7 @@ const PowerStateHandler = (deviceProvider, pnub) => ({
           if (dev.commands && dev.commands.power) {
             pnub.publish(gateway, {
               type: 'MQTT',
-              payload: {
-                topic: dev.commands.power.replace('mqtt:', ''),
-                value: param,
-              },
+              payload: factory.createPowerSwitchPayload(dev, param),
             });
             resolve();
           } else {