+++ src/events/handlers/internal/api/powerStateHandler.spec.js
@@ -5,16 +5,21 @@ import PowerStateHandler from './powerStateHandler';
 
 import helper from '../../processor_tests_helper.spec';
 helper('./powerStateProcessor');
+import payloadFactory from '../../factory/payloadFactory';
 
 import { DevicesProvider } from '../../../../data/mongodb/index';
 
 chai.should();
 const expect = chai.expect;
 
-describe('LwtHandler', () => {
+describe('Power Handler', () => {
   let subject;
   let deviceProvider;
   let pnub;
+  let createPowerPayload;
+  const fakePayload = {
+    topic: 'cmnd/test/POWER',
+  };
 
   beforeEach(() => {
     const db = {
@@ -24,6 +29,8 @@ describe('LwtHandler', () => {
     deviceProvider = DevicesProvider(db);
     pnub = {};
     subject = new PowerStateHandler(deviceProvider, pnub);
+
+    createPowerPayload = sinon.stub(payloadFactory.prototype, 'createPowerSwitchPayload').returns(fakePayload);
   });
 
   it('should publish enclosed mqtt message on PubNub', (done) => {
@@ -51,11 +58,15 @@ describe('LwtHandler', () => {
     subject.process(event)
       .then(() => {
         pnub.publish
-          .calledWith(sinon.match.any, sinon.match({ payload: { topic: 'cmnd/test6/POWER' } }))
+          .calledWith(sinon.match.any, sinon.match({ payload: fakePayload }))
           .should.be.true;
         done();
       })
       .catch(err => done(err));
+  });
+
+  afterEach(() => {
+    createPowerPayload.restore();
   });
 
 });