+++ src/events/mapper/toGateway/s6fresnel/S6FresnelPowerCommandToGateway.js
@@ -1,0 +1,6 @@
+const S6FresnelPowerToGateway = (dev, param) => ({
+  topic: dev.commands.power.replace('mqtt:', ''),
+  value: param,
+});
+
+export default S6FresnelPowerToGateway;