+++ src/events/mapper/toGateway/sonoff/SONPowerCommandToGateway.js
@@ -1,0 +1,6 @@
+const SONPowerToGateway = (dev, param) => ({
+  topic: dev.commands.power.replace('mqtt:', ''),
+  value: param,
+});
+
+export default SONPowerToGateway;